Fix compiler warnings

+++ src/alge/cs_sles_petsc.c
@@ -962,8 +962,7 @@ cs_sles_petsc_setup(void               *context,
       bft_error(__FILE__, __LINE__, 0,
                 _("Matrix type %s with block size %d for system \"%s\"\n"
                   "is not usable by PETSc."),
-                cs_matrix_get_type_name(a), db_size,
-                name);
+                cs_matrix_get_type_name(a), (int)db_size, name);
 
     /* Now preallocate matrix */
